We now calculate the width of the shortcut and its qualifiers and create a menu that is wide enough to accommodate everything.
The system hotkey file is sought by trimming the locale of '.' characters
and then '_' characters. So a locale like
en_AU.UTF-8
will look for hotkeys files successively:
en_AU.UTF-8
en_AU
en

Though maybe that should be a stack and we seek in reverse order to allow sub-languages
to overwrite the base. Or maybe that would itself be over engineering.

I only have an english hotkey file for now :/