Win32 build fix for encoding.
For whatever reason it seems that iconv is returning its result in
reverse byte order. It doesn't seem to matter if BE or LE is
stipulated at iconv_open time either, the result is always in the
wrong order, so a byte swap in the Enc() function might be the best
short term solution.

Washed down the patch from my build to this smaller change, hopefully it
has all the goodness